requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
requests==2.31.0
pydantic==2.5.0
python-multipart==0.0.6
pillow==10.1.0
python-dotenv==1.0.0
alembic==1.13.0
prometheus-client==0.19.0
schedule==1.2.0
openai==1.3.0
websockets==12.0
mcp==1.0.0